feat: add rendering of dynamic sections
