+++ apps/nextjs/src/app/[locale]/boards/_types.ts
@@ -7,5 +7,6 @@ export type Item = Section["items"][number];
 
 export type CategorySection = Extract<Section, { kind: "category" }>;
 export type EmptySection = Extract<Section, { kind: "empty" }>;
+export type DynamicSection = Extract<Section, { kind: "dynamic" }>;
 
 export type ItemOfKind<TKind extends WidgetKind> = Extract<Item, { kind: TKind }>;
