+++ apps/nextjs/src/components/board/sections/use-section-items.ts
@@ -4,7 +4,8 @@ import { useRequiredBoard } from "~/app/[locale]/boards/(content)/_context";
 export const useSectionItems = (section: Section) => {
   const board = useRequiredBoard();
   const innerSections = board.sections.filter(
-    (innerSection) => innerSection.kind === "dynamic" && innerSection.parentSectionId === section.id,
+    (innerSection): innerSection is Exclude<Section, { kind: "category" } | { kind: "empty" }> =>
+      innerSection.kind === "dynamic" && innerSection.parentSectionId === section.id,
   );
 
   return {
